Lock msm_create and mock the name info test_settings
